+++ src/font.h
@@ -1,3 +1,6 @@
+// Copyright (c) F4HWN Armel. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 const uint8_t rounded_led_board10pt7bBitmaps[] PROGMEM = {
     0x00, 0xFF, 0x8F, 0xF8, 0xFF, 0x8F, 0xF8, 0xFF, 0x80, 0x00, 0xFF, 0x80,
     0xE0, 0xFC, 0x1F, 0x83, 0x80, 0x0E, 0x0F, 0xC1, 0xF8, 0x38, 0x0E, 0x0E,
